pyproject.toml
[project]
name = "bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiogram>=3.22.0",
    "fastapi[standard]>=0.118.0",
    "python-decouple>=3.8",
    "requests>=2.32.5",
    "uvicorn>=0.37.0",
]
